refactor: Course 도메인 패키지 구조 정리 - CourseEdit, CourseState를 domain.model로 이동 - entity에는 JPA 엔티티(Course)만 유지
