+++ src/main/java/com/wanted/growthmate/learning/course/domain/model/CourseState.java
@@ -1,4 +1,4 @@
-package com.wanted.growthmate.learning.course.domain;
+package com.wanted.growthmate.learning.course.domain.model;
 
 public enum CourseState {
     DRAFT, //임시저장
